Actually fixed it this time
