+++ core/src/main/scala-2.13/org/bitcoins/core/util/MapWrapper.scala
@@ -1,12 +1,5 @@
 package org.bitcoins.core.util
 
-trait SeqWrapper[+T] extends IndexedSeq[T] {
-  protected def wrapped: IndexedSeq[T]
-  override def iterator: Iterator[T] = wrapped.iterator
-  override def length: Int = wrapped.length
-  override def apply(idx: Int): T = wrapped(idx)
-}
-
 trait MapWrapper[K, +T] extends Map[K, T] {
   protected def wrapped: Map[K, T]
   override def +[B1 >: T](kv: (K, B1)): Map[K, B1] = wrapped.+(kv)
@@ -14,4 +7,5 @@ trait MapWrapper[K, +T] extends Map[K, T] {
   override def iterator: Iterator[(K, T)] = wrapped.iterator
   override def updated[V1 >: T](key: K, value: V1): Map[K, V1] =
     wrapped.updated(key, value)
+  override def removed(key: K): Map[K, T] = wrapped.removed(key)
 }
