+++ dlc/src/main/scala/org/bitcoins/dlc/DLCMessage.scala
@@ -77,14 +77,6 @@ object DLCMessage {
       with MapWrapper[Sha256DigestBE, Satoshis] {
     override def wrapped: Map[Sha256DigestBE, Satoshis] = outcomeValueMap
 
-    def removed(key: Sha256DigestBE): ContractInfo = {
-      ContractInfo(outcomeValueMap - key)
-    }
-
-    def -(key: Sha256DigestBE): ContractInfo = {
-      ContractInfo(outcomeValueMap - key)
-    }
-
     override def bytes: ByteVector = {
       outcomeValueMap.foldLeft(ByteVector.empty) {
         case (vec, (digest, sats)) => vec ++ digest.bytes ++ sats.bytes
